==> 09_rf_binomial-response_all.R <==
# Random Forest predicting EWE or not based on adjusted fuel/topography/human factors

library(dplyr)
library(sf)
library(data.table)
library(spatialRF)

dir.create(file.path("data", "out", "rf"), recursive = TRUE, showWarnings = FALSE)

# system2(command = "aws", args = "s3 cp s3://california-megafires/data/out/fired_daily_ca_response-vars.csv  data/out/fired_daily_ca_response-vars.csv", stdout = TRUE)  
# 
# system2(command = "aws", args = "s3 cp s3://california-megafires/data/out/fired_events_ca_epsg3310_2003-2020.gpkg  data/out/fired_events_ca_epsg3310_2003-2020.gpkg", stdout = TRUE)  
# 
# system2(command = "aws", args = "s3 cp s3://california-megafires/tables/driver-descriptions.csv  tables/driver-descriptions.csv", stdout = TRUE)

biome_shortnames <- c("tcf", "mfws", "tgss", "dxs")

# Nonspatial binomial response
# version 1 of the binomial response model predicts whether daily area of increase is in top 95th percentile of daily area of increase
# version 2 of the binomial response model predicts whether daily area of increase is in top 95th percentile of daily area of increase as a regression problem
# version 3 of the binomial response model predicts whether daily AOI is in top 95th percentile, but only uses the first time that happens for each fire
# using "adjusted" variables, some of which are scale-dependent regardless of fire process
# version 4 is just like version 3 except we drop pdsi_z from the predictor list because of some anomolous values
# version 5 includes ALL the variables from the 'randomly placed FIRED perimeters' process
# version 6 combines some landform data (valley + narrow valley = valleys)
# version 7 fixes an issue where non-ewe days of an event after the first ewe event were included, while all ewe days after the first were excluded; back to just using all days from each fire
# version 8 adds event day
# version 9 uses the cleaner version of drivers data

# Nonspatial binomial response, just sqrt_aoi_tm1 as predictor
# version 1 of the binomial response model predicts whether daily area of increase is in top 95th percentile of daily area of increase
# version 2 of the binomial response model predicts whether daily area of increase is in top 95th percentile of daily area of increase as a regression problem
# version 3 of the binomial response model predicts whether daily AOI is in top 95th percentile, but only uses the first time that happens for each fire
nonspatial_version <- "v9"
nonspatial_simple_version <- "v3"
drivers_version <- "v6"

# For defining "ewe" or not, what is the percentage threshold? E.g., 0.95 means an "ewe" is in the top
# 5th percentile for daily area of increase
pct_threshold <- 0.95

for (counter in 1:4) {
  biome_shortname <- biome_shortnames[counter]
  
  # system2(command = "aws", args = paste0("s3 cp s3://california-megafires/data/out/analysis-ready/FIRED-daily-scale-drivers_california_", biome_shortname, "_v5.csv  data/out/analysis-ready/FIRED-daily-scale-drivers_california_", biome_shortname, "_", drivers_version, ".csv"), stdout = TRUE)  
  
  fired_daily_response <- 
    data.table::fread(input = "data/out/fired_daily_ca_response-vars.csv")
  
  driver_descriptions <- read.csv(file = "tables/driver-descriptions.csv")
  
  target_event_ids <-
    sf::read_sf("data/out/fired_events_ca_epsg3310_2003-2020.gpkg") %>% 
    dplyr::mutate(area_ha = as.numeric(sf::st_area(.)) / 10000) %>% 
    dplyr::filter(area_ha >= 121.406) %>% 
    dplyr::pull(id)
  
  fires <- 
    data.table::fread(paste0("data/out/analysis-ready/FIRED-daily-scale-drivers_california_", biome_shortname, "_", drivers_version, ".csv")) %>% 
    dplyr::filter(id %in% target_event_ids) %>% 
    dplyr::select(-max_wind_speed, -min_wind_speed, -max_rh, -min_rh, -max_temp, -min_temp, -max_soil_water, -min_soil_water, -max_vpd, -min_vpd,
                  -cumu_count, -cumu_area_ha,
                  -biggest_poly_area_ha, -x_3310, -y_3310, -biggest_poly_frac,
                  -ends_with("rtma"), -ends_with("rtma_pct"),
                  -bi, -erc, -fm100, -fm1000, -pdsi, -starts_with("spi"), -starts_with("eddi"),
                  -starts_with("raw"),
    ) %>% 
    dplyr::left_join(fired_daily_response) %>% 
    # dplyr::mutate(n_afd = ifelse(is.na(n_afd), yes = 0, no = n_afd),
    #               frp90 = ifelse(is.na(frp90), yes = 0, no = frp90),
    #               afd_per_ha = ifelse(is.na(afd_per_ha), yes = 0, no = afd_per_ha)) %>% 
    dplyr::mutate(sqrt_aoi_tm1 = sqrt(daily_area_tminus1_ha))
  
  idx <- substr(names(fires), start = 1, stop = 4) == "adj_"
  names(fires)[idx] <- substr(names(fires), start = 5, stop = nchar(names(fires)))[idx]
  
  # Defining "ewe" as whether daily area of increase was >95th percentile for the biome
  # 95th percentile growth for temperate conifer forests is 1683.266 ha per day
  # 95th percentile growth for Mediterranean Forest, Woodland & Scrub is 3341.724 ha per day
  # 95th percentile growth for temperate grasslands, savannas & shrublands is 195.0442 ha per day
  # 95th percentile growth for desert & xeric shrublands is 1335.305 ha per day
  fires <-
    fires  %>% 
    dplyr::mutate(area_log10_pct = ecdf(area_log10)(area_log10),
                  ewe = ifelse(area_log10_pct >= pct_threshold, yes = 1, no = 0)) %>% 
    dplyr::select(id, did, ewe, date, x_biggest_poly_3310, y_biggest_poly_3310, 
                  area_log10, change_diversity,
                  upper_slopes, lower_slopes, flat, barriers_to_spread, valleys,
                  starts_with("csp_ergo_landforms"), elevation, friction, 
                  friction_walking_only, landcover_diversity, landform_diversity, 
                  starts_with("lcms_change"), starts_with("lcms_landcover"), 
                  ndvi, road_density_mpha, rumple_index, 
                  valleys, veg_structure_rumple, npl,  
                  concurrent_fires, wind_anisotropy, wind_terrain_anisotropy, 
                  wind_terrain_alignment, max_wind_speed_pct, min_wind_speed_pct, 
                  max_rh_pct, min_rh_pct, max_temp_pct, min_temp_pct, 
                  max_soil_water_pct, min_soil_water_pct, max_vpd_pct, 
                  min_vpd_pct, starts_with("spei"), pdsi_z, erc_pct, bi_pct, 
                  fm100_pct, fm1000_pct, 
                  sqrt_aoi_tm1, cum_area_ha_tminus1, event_day) %>% 
    as.data.frame()
  
  # fires <-
  #   fires %>%
  #   dplyr::filter(ewe == 1) %>%
  #   dplyr::group_by(id) %>%
  #   dplyr::arrange(date) %>%
  #   dplyr::slice(1) %>%
  #   dplyr::ungroup() %>%
  #   rbind(fires[fires$ewe == 0, ]) %>%
  #   as.data.frame()
  
  fires <- 
    fires %>% 
    dplyr::rename(peak_ridge_warm = csp_ergo_landforms_11,
                  peak_ridge = csp_ergo_landforms_12,
                  peak_ridge_cool = csp_ergo_landforms_13,
                  mountain_divide = csp_ergo_landforms_14,
                  cliff = csp_ergo_landforms_15,
                  upper_slope_warm = csp_ergo_landforms_21,
                  upper_slope = csp_ergo_landforms_22,
                  upper_slope_cool = csp_ergo_landforms_23,
                  upper_slope_flat = csp_ergo_landforms_24,
                  lower_slope_warm = csp_ergo_landforms_31,
                  lower_slope = csp_ergo_landforms_32,
                  lower_slope_cool = csp_ergo_landforms_33,
                  lower_slope_flat = csp_ergo_landforms_34,
                  valley = csp_ergo_landforms_41,
                  valley_narrow = csp_ergo_landforms_42) %>% 
    dplyr::rename(trees = lcms_landcover_01,
                  shrubs_trees_mix = lcms_landcover_03,
                  grass_forbs_herb_trees_mix = lcms_landcover_04,
                  barren_trees_mix = lcms_landcover_05,
                  shrubs = lcms_landcover_07,
                  grass_forb_herb_shrub_mix = lcms_landcover_08,
                  barren_shrub_mix = lcms_landcover_09,
                  grass_forb_herb = lcms_landcover_10,
                  barren_grass_forb_herb_mix = lcms_landcover_11,
                  barren = lcms_landcover_12) %>% 
    dplyr::rename(fuel_stable = lcms_change_01,
                  fuel_slow_loss = lcms_change_02,
                  fuel_fast_loss = lcms_change_03,
                  fuel_gain = lcms_change_04) %>% 
    as.data.frame()
  
  
  # version 9
  human_drivers <- c("npl", "concurrent_fires", "friction_walking_only", "road_density_mpha")
  topography_drivers <- c("elevation", "rumple_index", "barriers_to_spread", "upper_slopes", "lower_slopes", "flat", "valleys", "landform_diversity")
  weather_drivers <- c("wind_anisotropy", "max_wind_speed_pct", "min_wind_speed_pct",
                       "max_temp_pct", "min_temp_pct",
                       "max_rh_pct", "min_rh_pct", "max_vpd_pct", "min_vpd_pct", "max_soil_water_pct", "min_soil_water_pct",
                       "spei14d", "spei30d", "spei90d", "spei180d", "spei270d", "spei1y", "spei2y", "spei5y", "fm100_pct", "fm1000_pct")
  fuel_drivers <- c("ndvi", "fuel_slow_loss", "fuel_fast_loss", "fuel_gain", "veg_structure_rumple", "landcover_diversity")
  interacting_drivers <- c("wind_terrain_anisotropy", "wind_terrain_alignment", "bi_pct", "erc_pct")
  
  predictor.variable.names <- c(human_drivers, topography_drivers, weather_drivers, fuel_drivers, interacting_drivers, "sqrt_aoi_tm1", "event_day")
  
  # version 8 (combined some landform categories)
  # human_drivers <- c("npl", "concurrent_fires", "friction_walking_only", "road_density_mpha")
  # topography_drivers <- c("elevation", "rumple_index", "barriers_to_spread", "upper_slopes", "lower_slopes", "flat", "valleys", "landform_diversity")
  # weather_drivers <- c("wind_anisotropy", "max_wind_speed_pct", "min_wind_speed_pct",
  #                      "max_temp_pct", "min_temp_pct",
  #                      "max_rh_pct", "min_rh_pct", "max_vpd_pct", "min_vpd_pct", "max_soil_water_pct", "min_soil_water_pct",
  #                      "spei14d", "spei30d", "spei90d", "spei180d", "spei270d", "spei1y", "spei2y", "spei5y", "fm100_pct", "fm1000_pct")
  # fuel_drivers <- c("ndvi", "fuel_slow_loss", "fuel_fast_loss", "fuel_gain", "veg_structure_rumple", "landcover_diversity")
  # interacting_drivers <- c("wind_terrain_anisotropy", "wind_terrain_alignment", "bi_pct", "erc_pct")
  # 
  # predictor.variable.names <- c(human_drivers, topography_drivers, weather_drivers, fuel_drivers, interacting_drivers, "sqrt_aoi_tm1", "event_day")
  
  # version 6 and 7
  # human_drivers <- c("npl", "concurrent_fires", "friction_walking_only", "road_density_mpha")
  # topography_drivers <- c("elevation", "rumple_index", "barriers_to_spread", "upper_slopes", "lower_slopes", "flat", "valleys", "landform_diversity")
  # weather_drivers <- c("wind_anisotropy", "max_wind_speed_pct", "min_wind_speed_pct",
  #                      "max_temp_pct", "min_temp_pct",
  #                      "max_rh_pct", "min_rh_pct", "max_vpd_pct", "min_vpd_pct", "max_soil_water_pct", "min_soil_water_pct",
  #                      "spei14d", "spei30d", "spei90d", "spei180d", "spei270d", "spei1y", "spei2y", "spei5y", "fm100_pct", "fm1000_pct")
  # fuel_drivers <- c("ndvi", "fuel_slow_loss", "fuel_fast_loss", "fuel_gain", "veg_structure_rumple", "landcover_diversity")
  # interacting_drivers <- c("wind_terrain_anisotropy", "wind_terrain_alignment", "bi_pct", "erc_pct")
  # 
  # predictor.variable.names <- c(human_drivers, topography_drivers, weather_drivers, fuel_drivers, interacting_drivers, "sqrt_aoi_tm1")
  # 
  # version 5
  # human_drivers <- c("npl", "concurrent_fires", "friction_walking_only", "road_density_mpha")
  # topography_drivers <- c("elevation", "rumple_index", "peak_ridge_warm", "peak_ridge", "peak_ridge_cool", "mountain_divide", "cliff", "upper_slope_warm", "upper_slope", "upper_slope_cool", "lower_slope_warm", "lower_slope", "lower_slope_cool", "valley", "valley_narrow", "landform_diversity")
  # weather_drivers <- c("wind_anisotropy", "max_wind_speed_pct", "min_wind_speed_pct",
  #                      "max_temp_pct", "min_temp_pct",
  #                      "max_rh_pct", "min_rh_pct", "max_vpd_pct", "min_vpd_pct", "max_soil_water_pct", "min_soil_water_pct",
  #                      "spei14d", "spei30d", "spei90d", "spei180d", "spei270d", "spei1y", "spei2y", "spei5y", "fm100_pct", "fm1000_pct")
  # fuel_drivers <- c("ndvi", "fuel_slow_loss", "fuel_fast_loss", "fuel_gain", "veg_structure_rumple", "trees", "shrubs_trees_mix", "grass_forbs_herb_trees_mix", "barren_trees_mix", "shrubs", "grass_forb_herb_shrub_mix", "barren_shrub_mix", "grass_forb_herb", "barren_grass_forb_herb_mix", "barren", "landcover_diversity")
  # interacting_drivers <- c("wind_terrain_anisotropy", "wind_terrain_alignment", "bi_pct", "erc_pct")
  
  
  # No NAs
  apply(fires[, predictor.variable.names], MARGIN = 2, FUN = function(x) return(sum(is.na(x))))
  
  # Remove these columns that have more NA's than most so we opt to drop them instead of dropping the rows
  na_cols <- colnames(fires[, predictor.variable.names])[which(apply(fires[, predictor.variable.names], 2, function(x) return(sum(is.na(x)))) > 50)]
  
  predictor.variable.names <- predictor.variable.names[!(predictor.variable.names %in% na_cols)]
  fires <- 
    fires %>% 
    dplyr::select(-all_of(na_cols))
  
  # Now drop all fires that have an NA in any column
  bad_fires <- unique(fires[!complete.cases(fires[, predictor.variable.names]), "id"])
  
  fires <- fires[!(fires$id %in% bad_fires), ]
  
  # no columnns with 0 variance (rounded to 4 decimal places)
  zero_variance_columns <- colnames(fires[, predictor.variable.names])[round(apply(fires[, predictor.variable.names], MARGIN = 2, FUN = var), 4) == 0]
  
  predictor.variable.names <- predictor.variable.names[!(predictor.variable.names %in% zero_variance_columns)]
  
  fires <- 
    fires %>% 
    dplyr::select(-all_of(zero_variance_columns))
  
  # No NaN or Inf when scaling
  sum(apply(scale(fires[, predictor.variable.names]), 2, is.nan))
  sum(apply(scale(fires[, predictor.variable.names]), 2, is.infinite))
  
  sort(cor(fires[, predictor.variable.names])[, "erc_pct"])
  
  # Reduce collinearity in the predictors
  preference.order <- c(
    "npl", 
    "rumple_index", "elevation",
    "max_wind_speed_pct", "min_wind_speed_pct",
    "max_vpd_pct", "min_vpd_pct",
    "fm100_pct", "erc_pct", "spei1y",
    "ndvi", "veg_structure_rumple", "landcover_diversity", "change_diversity",
    "wind_terrain_alignment"
  )
  
  predictor.variable.names_reduced <- spatialRF::auto_cor(
    x = fires[, predictor.variable.names],
    cor.threshold = 0.75,
    preference.order = preference.order
  ) %>% 
    spatialRF::auto_vif(
      vif.threshold = 5,
      preference.order = preference.order
    )
  
  ##### ---- SET UP DATA SUBSETS WITH THEIR XY MATRICES
  data <- fires
  distance_thresholds = c(0, 1000, 2000, 4000, 8000, 16000, 32000, 64000)
  
  random_seed <- 2203
  xy <- data[, c("x_biggest_poly_3310", "y_biggest_poly_3310")] %>% setNames(c("x", "y"))
  
  data_sf <-
    data %>%
    sf::st_as_sf(coords = c("x_biggest_poly_3310", "y_biggest_poly_3310"), crs = 3310, remove = FALSE)
  
  dist_mat <-
    sf::st_distance(x = data_sf,
                    y = data_sf) %>%
    units::drop_units()
  
  ### ----- FEATURE ENGINEERING
  
  # (start_time <- Sys.time())
  # ftr_eng <- spatialRF::the_feature_engineer(data = data,
  #                                            dependent.variable.name = "ewe",
  #                                            predictor.variable.names = predictor.variable.names_reduced,
  #                                            xy = xy)
  # (end_time <- Sys.time())
  # (difftime(time1 = end_time, time2 = start_time, units = "mins"))
  # Fitting and evaluating a model without interactions.
  # Testing 55 candidate interactions.
  # No promising interactions found. 
  # 
  # > (end_time <- Sys.time())
  # [1] "2022-07-27 13:55:23 MDT"
  # > (difftime(time1 = end_time, time2 = start_time, units = "mins"))
  # Time difference of 8.721521 mins
  
  (start_time <- Sys.time())
  biome_nonspatial <- spatialRF::rf(
    data = data,
    dependent.variable.name = "ewe",
    predictor.variable.names = predictor.variable.names_reduced,
    distance.matrix = dist_mat,
    distance.thresholds = distance_thresholds,
    xy = xy, #not needed by rf, but other functions read it from the model
    seed = random_seed,
    verbose = TRUE
  ) 
  
  biome_nonspatial <-
    spatialRF::rf_evaluate(
      model = biome_nonspatial,
      xy = xy,                  #data coordinates
      repetitions = 30,         #number of spatial folds
      training.fraction = 0.75, #training data fraction on each fold
      metrics = c("r.squared", "rmse", "auc"),
      seed = random_seed,
      verbose = TRUE
    )
  (end_time <- Sys.time())
  (difftime(time1 = end_time, time2 = start_time, units = "mins"))
  
  biome_nonspatial_import <- spatialRF::rf_importance(model = biome_nonspatial, xy = xy, metric = "rmse")
  
  readr::write_rds(x = biome_nonspatial, file = file.path("data", "out", "rf", paste0("rf_", biome_shortname, "_binomial-response-95th-pct-ewe_nonspatial_", nonspatial_version, ".rds")))
  readr::write_rds(x = biome_nonspatial_import, file = file.path("data", "out", "rf", paste0("rf_", biome_shortname, "_binomial-response-95th-pct-ewe_nonspatial-model-transferability_", nonspatial_version, ".rds")))
  
  # Repeat to get a distribution of importance values
  biome_nonspatial_repeat <- spatialRF::rf_repeat(
    model = biome_nonspatial,
    repetitions = 30,
    seed = random_seed,
    verbose = TRUE
  )
  
  readr::write_rds(x = biome_nonspatial_repeat, file = file.path("data", "out", "rf", paste0("rf_", biome_shortname, "_binomial-response-95th-pct-ewe_nonspatial-repeat_", nonspatial_version, ".rds")))
  
}


# Fit and write to disk the simple model, using just sqrt_aoi_tm1 as the only predictor
for (j in seq_along(biome_shortnames)) {
  
  biome_shortname <- biome_shortnames[j]
  
  fired_daily_response <- 
    data.table::fread(input = "data/out/fired_daily_ca_response-vars.csv")
  
  driver_descriptions <- read.csv(file = "tables/driver-descriptions.csv")
  
  target_event_ids <-
    sf::read_sf("data/out/fired_events_ca_epsg3310_2003-2020.gpkg") %>% 
    dplyr::mutate(area_ha = as.numeric(sf::st_area(.)) / 10000) %>% 
    dplyr::filter(area_ha >= 121.406) %>% 
    dplyr::pull(id)
  
  fires <- 
    data.table::fread(paste0("data/out/analysis-ready/FIRED-daily-scale-drivers_california_", biome_shortname, "_", drivers_version, ".csv")) %>% 
    dplyr::filter(id %in% target_event_ids) %>% 
    dplyr::select(-max_wind_speed, -min_wind_speed, -max_rh, -min_rh, -max_temp, -min_temp, -max_soil_water, -min_soil_water, -max_vpd, -min_vpd,
                  -cumu_count, -cumu_area_ha,
                  -biggest_poly_area_ha, -x_3310, -y_3310, -biggest_poly_frac,
                  -ends_with("rtma"), -ends_with("rtma_pct"),
                  -bi, -erc, -fm100, -fm1000, -pdsi, -starts_with("spi"), -starts_with("eddi"),
                  -starts_with("raw"),
    ) %>% 
    dplyr::left_join(fired_daily_response) %>% 
    # dplyr::mutate(n_afd = ifelse(is.na(n_afd), yes = 0, no = n_afd),
    #               frp90 = ifelse(is.na(frp90), yes = 0, no = frp90),
    #               afd_per_ha = ifelse(is.na(afd_per_ha), yes = 0, no = afd_per_ha)) %>% 
    dplyr::mutate(sqrt_aoi_tm1 = sqrt(daily_area_tminus1_ha)) %>% 
    dplyr::mutate(area_log10_pct = ecdf(area_log10)(area_log10),
                  ewe = ifelse(area_log10_pct >= pct_threshold, yes = 1, no = 0)) %>% 
    as.data.frame()
  
  idx <- substr(names(fires), start = 1, stop = 4) == "adj_"
  names(fires)[idx] <- substr(names(fires), start = 5, stop = nchar(names(fires)))[idx]
  
  data <- fires
  distance_thresholds = c(0, 1000, 2000, 4000, 8000, 16000, 32000, 64000)
  
  random_seed <- 2203
  xy <- data[, c("x_biggest_poly_3310", "y_biggest_poly_3310")] %>% setNames(c("x", "y"))
  
  data_sf <-
    data %>%
    sf::st_as_sf(coords = c("x_biggest_poly_3310", "y_biggest_poly_3310"), crs = 3310, remove = FALSE)
  
  dist_mat <-
    sf::st_distance(x = data_sf,
                    y = data_sf) %>%
    units::drop_units()
  
  (start_time <- Sys.time())
  biome_nonspatial_simple <- spatialRF::rf(
    data = data,
    dependent.variable.name = "ewe",
    predictor.variable.names = "sqrt_aoi_tm1",
    distance.matrix = dist_mat,
    distance.thresholds = distance_thresholds,
    xy = xy, #not needed by rf, but other functions read it from the model
    seed = random_seed,
    verbose = TRUE
  )
  (end_time <- Sys.time())
  (difftime(time1 = end_time, time2 = start_time, units = "mins"))
  
  biome_nonspatial_simple <- spatialRF::rf_evaluate(
    model = biome_nonspatial_simple,
    xy = xy,                  #data coordinates
    repetitions = 30,         #number of spatial folds
    training.fraction = 0.75, #training data fraction on each fold
    metrics = c("r.squared", "rmse", "auc"),
    seed = random_seed,
    verbose = TRUE
  )
  
  readr::write_rds(x = biome_nonspatial_simple, file = file.path("data", "out", "rf", paste0("rf_", biome_shortname, "_binomial-response-95th-pct-ewe_nonspatial-simple_", nonspatial_simple_version, ".rds")))
}
